+++ BWidgets/HRangeScrollBar.hpp
@@ -477,8 +477,6 @@ inline void HRangeScrollBar::update ()
 		button2.moveTo (x + (w - h) * (step_.second >= 0.0 ? rv.second : 1.0 - rv.second), y);
 		button2.resize (h, h);
 
-		const BStyles::Color bgColor = getBgColors()[getStatus()];
-		const BStyles::Color fgColor = getFgColors()[getStatus()];
 		cairo_t* cr1 = cairo_create (button1.cairoSurface());
 		drawKnob (cr1, 0.5 * h, 0.5 * h, BWIDGETS_DEFAULT_SCROLLBAR_KNOB_REL_SIZE * h, 0.0, BWIDGETS_DEFAULT_SCROLLBAR_KNOB_COLOR, BWIDGETS_DEFAULT_SCROLLBAR_KNOB_COLOR);
 		cairo_destroy (cr1);