+++ BWidgets/VScale.hpp
@@ -300,8 +300,6 @@ inline void VScale::draw (const BUtilities::Area<>& area)
 			cairo_clip (cr);
 
 			const double rval = getRatioFromValue (getValue());
-			const BStyles::Color fgColor = getFgColors()[getStatus()];
-			const BStyles::Color bgColor = getBgColors()[getStatus()];
 
 			if (step_ >= 0.0) 
 			{