+++ BWidgets/VScrollBar.hpp
@@ -277,8 +277,6 @@ inline void VScrollBar::draw (const BUtilities::Area<>& area)
 
 			const double rval1 = (step_ >= 0.0 ? getRatioFromValue (getValue()) : 1.0 - getRatioFromValue (getValue() + vsize_));
 			const double rval2 = (step_ >= 0.0 ? getRatioFromValue (getValue() + vsize_) : 1.0 - getRatioFromValue (getValue()));
-			const BStyles::Color fgColor = getFgColors()[getStatus()];
-			const BStyles::Color bgColor = getBgColors()[getStatus()];
 			drawVBar
 			(
 				cr,