+++ BWidgets/VSwitch.hpp
@@ -289,8 +289,6 @@ inline void VSwitch::draw (const BUtilities::Area<>& area)
 			const double xc = x0 + 0.5 * weff;
 			const double yc = y0 + heff - radius - max * (heff - 2.0 * radius);
 
-			const BStyles::Color fgColor = getFgColors()[getStatus()];
-			const BStyles::Color bgColor = getBgColors()[getStatus()];
 			drawVBar(cr, x0 + 0.5 * (1.0 - BWIDGETS_DEFAULT_SWITCH_BAR_REL_SIZE) * weff, y0, BWIDGETS_DEFAULT_SWITCH_BAR_REL_SIZE * weff, heff, 1.0, 1.0 - max, BWIDGETS_DEFAULT_SWITCH_BAR_FGCOLOR, BWIDGETS_DEFAULT_SWITCH_BAR_BGCOLOR);
 			drawKnob(cr, xc, yc, radius, 1.0, BWIDGETS_DEFAULT_SWITCH_KNOB_COLOR, BWIDGETS_DEFAULT_SWITCH_KNOB_COLOR);
 		}